Make ColorCenter coordinates int
A simple one-liner casting the coordinates in the return of _findColorCenter(self, color) to ints. 

This fixes a problem I encountered: If coordinates of start/end point are in between two pixels, e.g. X.5, incrementing/decrementing the coordinates by 1 would never land exactly on the end point if its coordinates are X.0, or vice versa.